// CSE_308/Spring-Files/CSE_308_Overdrive/overdrive_frontend/src/models/users/users.ts
export interface Users {
  // private String username;
	// private String password;
	// private String email;
	// private BsonType profilePic;
	// private String bio;
	// private List<String> followedSeries;
	// private List<String> producedSeries;
  username: string;
  password: string;
  email: string;
  profilePic: string;
  bio: string;
  editorPics: string[];
}